backend/requirements.txt
fastapi
uvicorn
openai
whisper
python-dotenv
